Lazy load booking calendar logic on interaction.
Extracted availability calculation logic from `booking-ui.js` to `js/booking-calendar.js`.
Modified `booking-ui.js` to dynamically import `js/booking-calendar.js` when needed.
Added click listeners to `index.html` to lazy load the calendar script when booking buttons are clicked, improving initial load performance.
