+++ booking-ui.js
@@ -100,40 +100,18 @@
         initialized = true;
 
         async function updateAvailableTimes() {
-            const selectedDate = dateInput ? dateInput.value : '';
-            if (!selectedDate || !timeSelect) return;
-
-            const selectedDoctor = doctorSelect ? doctorSelect.value : '';
-            const availability = await deps.loadAvailabilityData();
-            const bookedSlots = await deps.getBookedSlots(selectedDate, selectedDoctor);
-            const availableSlots = availability[selectedDate] || deps.getDefaultTimeSlots();
-            const isToday = selectedDate === new Date().toISOString().split('T')[0];
-            const nowMinutes = isToday ? new Date().getHours() * 60 + new Date().getMinutes() : -1;
-            const freeSlots = availableSlots.filter((slot) => {
-                if (bookedSlots.includes(slot)) return false;
-                if (isToday) {
-                    const [h, m] = slot.split(':').map(Number);
-                    if (h * 60 + m <= nowMinutes + 60) return false;
-                }
-                return true;
-            });
-
-            const currentValue = timeSelect.value;
-            timeSelect.innerHTML = '<option value="">Hora</option>';
-
-            if (freeSlots.length === 0) {
-                timeSelect.innerHTML += '<option value="" disabled>No hay horarios disponibles</option>';
-                deps.showToast(t('No hay horarios disponibles para esta fecha', 'No slots available for this date'), 'warning');
-                return;
+            try {
+                const module = await import('./js/booking-calendar.js');
+                await module.updateAvailableTimes(deps, {
+                    dateInput,
+                    timeSelect,
+                    doctorSelect,
+                    t
+                });
+            } catch (error) {
+                console.error('Failed to load booking-calendar.js', error);
+                deps.showToast(t('Error cargando calendario. Intenta nuevamente.', 'Error loading calendar. Please try again.'), 'error');
             }
-
-            freeSlots.forEach((time) => {
-                const option = document.createElement('option');
-                option.value = time;
-                option.textContent = time;
-                if (time === currentValue) option.selected = true;
-                timeSelect.appendChild(option);
-            });
         }
 
         serviceSelect.addEventListener('change', function () {
